encapsulate menu behaviour, improve a11y
