+++ packages/site-kit/src/lib/nav/Menu.svelte
@@ -1,11 +1,28 @@
 <script>
+	import { expoOut, quintOut } from 'svelte/easing';
 	import { afterNavigate } from '$app/navigation';
-	import { click_outside, focus_outside } from '$lib/actions';
-	import { overlay_open } from '$lib/stores';
+	import { click_outside, focus_outside, trap } from '$lib/actions';
+	import { overlay_open, reduced_motion, theme } from '$lib/stores';
 	import Icon from '../components/Icon.svelte';
+	import NavContextMenu from './NavContextMenu.svelte';
+	import { page } from '$app/stores';
 
 	/** @type {boolean} */
 	export let open;
+
+	/** @type {import('../types').NavigationLink[]} */
+	export let links;
+
+	/** @type {import('../types').NavigationLink | undefined} */
+	let current_menu_view = undefined;
+
+	/** @type {NavContextMenu} */
+	let nav_context_instance;
+
+	let menu_height = 0;
+	let universal_menu_inner_height = 0;
+	let ready = false;
+	let show_context_menu = false;
 
 	function close() {
 		open = false;
@@ -13,8 +30,54 @@
 
 	afterNavigate(close);
 
+	/**
+	 * @param {HTMLElement} _
+	 * @param {(current: boolean) => void} fn
+	 */
+	function mounted(_, fn) {
+		// this is necessary to ensure that the menu-background height
+		// is applied without an animation
+		setTimeout(() => {
+			fn(true);
+		});
+
+		return {
+			destroy() {
+				fn(false);
+			}
+		};
+	}
+
+	/**
+	 * @param {HTMLElement} node
+	 * @param {{easing?: (t: number) => number, duration?: number }} [params]
+	 * @returns {import('svelte/transition').TransitionConfig}
+	 */
+	const slide = (node, { duration = 400, easing = expoOut } = {}) => {
+		const height = current_menu_view ? node.clientHeight : universal_menu_inner_height;
+
+		return {
+			css: (t, u) =>
+				$reduced_motion.current
+					? `opacity: ${t}`
+					: `transform: translate3d(0, ${height * u}px, 0) scale3d(${0.9 + 0.1 * t}, ${
+							0.9 + 0.1 * t
+					  }, 1)`,
+			easing,
+			duration
+		};
+	};
+
 	$: $overlay_open = open;
 </script>
+
+<svelte:window
+	on:keydown={(e) => {
+		if (e.key === 'Escape') {
+			close();
+		}
+	}}
+/>
 
 <div style="display: contents" use:click_outside={close} use:focus_outside={close}>
 	<button
@@ -22,14 +85,118 @@
 		aria-expanded={open}
 		class="menu-toggle"
 		class:open
-		on:click={() => (open = !open)}
+		on:click={() => {
+			if (open) {
+				close();
+			} else {
+				open = true;
+
+				const segment = $page.url.pathname.split('/')[1];
+				current_menu_view = links.find((link) => link.prefix === segment);
+				show_context_menu = !!current_menu_view;
+			}
+		}}
 	>
 		<Icon name={open ? 'close' : 'menu'} size="1em" />
 	</button>
 
 	{#if open}
-		<div class="menu">
-			<slot />
+		<div class="menu" use:trap>
+			<div class="mobile-main-menu" in:slide out:slide={{ duration: 500, easing: quintOut }}>
+				<div
+					class="menu-background"
+					class:dark={$theme.current === 'dark'}
+					class:ready
+					style:height={show_context_menu ? '100%' : `${universal_menu_inner_height}px`}
+					style:--background={show_context_menu ? 'var(--sk-back-3)' : null}
+					use:mounted={(mounted) => (ready = mounted)}
+				/>
+
+				<div
+					class="clip"
+					style:--height-difference="{menu_height - universal_menu_inner_height}px"
+					on:transitionstart={(e) => {
+						const target = /** @type {HTMLElement} */ (e.target);
+
+						if (!target?.classList.contains('viewport')) return;
+						if (e.propertyName !== 'transform') return;
+
+						// we need to apply a clip-path during the transition so that the contents
+						// are constrained to the menu background, but only while the transition
+						// is running, otherwise it prevents the contents from being scrolled
+						const a = 'calc(var(--height-difference) + 10px)';
+						const b = '10px';
+
+						const start = show_context_menu ? a : b;
+						const end = show_context_menu ? b : a;
+
+						const container = e.currentTarget;
+
+						container.style.clipPath = `polygon(0% ${start}, 100% ${start}, 100% 100%, 0% 100%)`;
+
+						setTimeout(() => {
+							container.style.clipPath = `polygon(0% ${end}, 100% ${end}, 100% 100%, 0% 100%)`;
+						}, 0);
+					}}
+					on:transitionend={(e) => {
+						const target = /** @type {HTMLElement} */ (e.target);
+
+						if (!target?.classList.contains('viewport')) return;
+						if (e.propertyName !== 'transform') return;
+
+						e.currentTarget.style.clipPath = '';
+					}}
+				>
+					<div
+						class="viewport"
+						class:reduced-motion={$reduced_motion.current}
+						class:offset={show_context_menu}
+						bind:clientHeight={menu_height}
+					>
+						<div class="universal" inert={show_context_menu}>
+							<div class="contents" bind:clientHeight={universal_menu_inner_height}>
+								{#each links as link}
+									<a href={link.pathname}>
+										{link.title}
+
+										{#if link.sections}
+											<button
+												class="related-menu-arrow"
+												on:click|preventDefault={() => {
+													current_menu_view = link;
+													show_context_menu = true;
+												}}
+											>
+												<Icon name="arrow-right-chevron" size="6rem" />
+											</button>
+										{/if}
+									</a>
+								{/each}
+
+								<slot />
+							</div>
+						</div>
+
+						<div class="context" inert={!show_context_menu}>
+							{#if current_menu_view}
+								<NavContextMenu
+									bind:this={nav_context_instance}
+									contents={current_menu_view.sections}
+								/>
+							{/if}
+						</div>
+
+						<button
+							class="back-button"
+							on:click={() => (show_context_menu = false)}
+							inert={!show_context_menu}
+						>
+							<Icon name="arrow-left" size=".6em" />
+							<span>Back to main menu</span>
+						</button>
+					</div>
+				</div>
+			</div>
 		</div>
 	{/if}
 </div>
@@ -61,5 +228,155 @@
 		gap: 1.5rem;
 
 		line-height: 1;
+	}
+
+	.menu-background {
+		position: absolute;
+		width: 100%;
+		left: 0;
+		bottom: 0;
+		height: 100%;
+		border-radius: 1rem 1rem 0 0;
+		background: var(--background, var(--sk-back-2));
+	}
+
+	.menu-background.ready {
+		transition: height 0.4s cubic-bezier(0.23, 1, 0.32, 1);
+	}
+
+	.menu-background.dark {
+		border-top: solid 1.1px hsla(0, 0%, 100%, 0.2);
+	}
+
+	.mobile-main-menu {
+		height: 100%;
+	}
+
+	.clip {
+		width: 100%;
+		height: 100%;
+		transition: clip-path 0.4s cubic-bezier(0.23, 1, 0.32, 1);
+	}
+
+	.viewport {
+		position: relative;
+		display: grid;
+		width: 200%;
+		height: 100%;
+		grid-template-columns: 50% 50%;
+		transition: transform 0.4s cubic-bezier(0.23, 1, 0.32, 1);
+		grid-auto-rows: 100%;
+	}
+
+	.viewport.reduced-motion {
+		transition: none;
+	}
+
+	.viewport.offset {
+		transform: translate3d(-50%, 0, 0);
+	}
+
+	.universal .contents {
+		position: absolute;
+		width: 50%;
+		bottom: 0;
+		padding: 1rem;
+		max-height: 70vh;
+		overflow-y: scroll;
+	}
+
+	.viewport > * {
+		overflow-y: auto;
+		transition: inherit;
+		transition-property: transform, opacity;
+	}
+
+	.context {
+		position: relative;
+		height: 100%;
+		padding-bottom: 2rem;
+	}
+
+	.back-button {
+		position: absolute;
+		bottom: 0;
+		right: 0;
+		z-index: 9;
+
+		display: flex;
+		align-items: center;
+		justify-content: start;
+		gap: 1rem;
+
+		font-size: 0.9em;
+		color: var(--sk-text-3);
+
+		background-color: var(--sk-back-2);
+
+		width: 50%;
+		height: 48px;
+		padding: 0 1.5rem;
+	}
+
+	.back-button :global(svg) {
+		transform: scale(0.8);
+	}
+
+	.viewport :global(a) {
+		position: relative;
+		display: block;
+		align-items: center;
+		padding: 0.3rem 0;
+		height: 4rem;
+		color: var(--sk-text-2);
+	}
+
+	.universal .contents,
+	.context,
+	.back-button {
+		pointer-events: all;
+	}
+
+	.universal :global(a) {
+		position: relative;
+	}
+
+	.universal .contents :global(a button) {
+		position: absolute;
+		right: 0;
+		top: 0;
+		width: 4rem;
+		height: 100%;
+	}
+
+	.viewport :global(a svg) {
+		stroke-width: 0;
+	}
+
+	.viewport :global(a) {
+		display: flex;
+		align-items: center;
+		border-radius: var(--sk-border-radius);
+		width: 100%;
+		height: 100%;
+		padding-left: 1rem;
+	}
+
+	.viewport :global(a[aria-current='true']) {
+		background-color: hsla(var(--sk-theme-1-hsl), 0.05);
+	}
+
+	.viewport :global(a:hover) {
+		border-radius: var(--sk-border-radius);
+
+		color: initial;
+		text-decoration: none;
+
+		background-color: var(--sk-back-4);
+	}
+
+	.viewport :global(a[aria-current='true']:hover) {
+		background-color: hsla(var(--sk-theme-1-hsl), 0.05);
+		color: var(--sk-theme-1);
 	}
 </style>
